Refactor (for better testability)
