+++ sdk/src/CMakeLists.txt
@@ -38,8 +38,10 @@ add_library(${TARGET_NAME}
     sdk/pubsub/MqttPubSubClient.cpp
     sdk/vdb/DataPointBatch.cpp
     sdk/vdb/IVehicleDataBrokerClient.cpp
-    sdk/vdb/grpc/KuksaValV2AsyncGrpcFacade.cpp
-    sdk/vdb/grpc/KuksaValV2Client.cpp
+    sdk/vdb/grpc/common/TypeConversions.cpp
+    sdk/vdb/grpc/kuksa_val_v2/BrokerAsyncGrpcFacade.cpp
+    sdk/vdb/grpc/kuksa_val_v2/BrokerClient.cpp
+    sdk/vdb/grpc/kuksa_val_v2/TypeConversions.cpp
 )
 
 include_directories(
