+++ sdk/src/sdk/vdb/grpc/kuksa_val_v2/BrokerAsyncGrpcFacade.cpp
@@ -14,18 +14,19 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "sdk/vdb/grpc/KuksaValV2AsyncGrpcFacade.h"
+#include "BrokerAsyncGrpcFacade.h"
+
 #include "sdk/Logger.h"
 #include "sdk/grpc/GrpcCall.h"
 
 #include <grpcpp/channel.h>
 
-namespace velocitas {
+namespace velocitas::kuksa_val_v2 {
 
-KuksaValV2AsyncGrpcFacade::KuksaValV2AsyncGrpcFacade(std::shared_ptr<grpc::Channel> channel)
+BrokerAsyncGrpcFacade::BrokerAsyncGrpcFacade(std::shared_ptr<grpc::Channel> channel)
     : m_stub{kuksa::val::v2::VAL::NewStub(channel)} {}
 
-void KuksaValV2AsyncGrpcFacade::GetValues(
+void BrokerAsyncGrpcFacade::GetValues(
     kuksa::val::v2::GetValuesRequest&&                                  request,
     std::function<void(const kuksa::val::v2::GetValuesResponse& reply)> replyHandler,
     std::function<void(const grpc::Status& status)>                     errorHandler) {
@@ -55,7 +56,7 @@ void KuksaValV2AsyncGrpcFacade::GetValues(
                                grpcResultHandler);
 }
 
-void KuksaValV2AsyncGrpcFacade::BatchActuate(
+void BrokerAsyncGrpcFacade::BatchActuate(
     kuksa::val::v2::BatchActuateRequest&&                                  request,
     std::function<void(const kuksa::val::v2::BatchActuateResponse& reply)> replyHandler,
     std::function<void(const grpc::Status& status)>                        errorHandler) {
@@ -86,7 +87,7 @@ void KuksaValV2AsyncGrpcFacade::BatchActuate(
                                   grpcResultHandler);
 }
 
-void KuksaValV2AsyncGrpcFacade::Subscribe(
+void BrokerAsyncGrpcFacade::Subscribe(
     kuksa::val::v2::SubscribeRequest&&                                  request,
     std::function<void(const kuksa::val::v2::SubscribeResponse& reply)> itemHandler,
     std::function<void(const grpc::Status& status)>                     errorHandler) {
@@ -115,4 +116,4 @@ void KuksaValV2AsyncGrpcFacade::Subscribe(
     callData->startCall();
 }
 
-} // namespace velocitas
+} // namespace velocitas::kuksa_val_v2
